+++ packages/gamut-labs/src/CurriculumCard/index.tsx
@@ -9,6 +9,8 @@ import { Subtitle, SubtitleProps } from './Subtitle';
 
 export type ProgressState = 'inProgress' | 'completed';
 
+export type FooterTextVariantType = 'enrolled' | 'inProgress';
+
 const cardHeight = 180;
 
 const cardStyles = {
@@ -27,6 +29,11 @@ export type CurriculumCardProps = SubtitleProps & {
   showProLogo?: boolean;
   tag?: string;
   tagColor?: TagColor;
+  showAltSubtitle?: boolean;
+  /**
+   * Changes In Progress card footer text from "Enrolled..." to "In Progress..."
+   */
+  footerTextVariant?: FooterTextVariantType;
 };
 
 export const CurriculumCard: React.FC<CurriculumCardProps> = ({
@@ -37,12 +44,13 @@ export const CurriculumCard: React.FC<CurriculumCardProps> = ({
   isFullSize = false,
   progressState,
   scope,
-  scopeCount,
   showProLogo,
   tag,
   tagColor,
   text,
   title,
+  showAltSubtitle = false,
+  footerTextVariant = 'enrolled',
 }) => {
   const boxVariant = progressState && cardStyles[progressState];
   const mode = progressState === 'completed' ? 'dark' : 'light';
@@ -73,8 +81,8 @@ export const CurriculumCard: React.FC<CurriculumCardProps> = ({
         {!progressState && (
           <Subtitle
             scope={scope}
-            scopeCount={scopeCount}
             difficulty={difficulty}
+            showAltSubtitle={showAltSubtitle}
           />
         )}
       </FlexBox>
@@ -88,6 +96,7 @@ export const CurriculumCard: React.FC<CurriculumCardProps> = ({
         progressState={progressState}
         tag={tag}
         tagColor={tagColor}
+        footerTextVariant={footerTextVariant}
       />
     </Card>
   );