+++ packages/gamut-labs/src/Notifications/__tests__/NotificationsContents-test.tsx
@@ -16,6 +16,44 @@ const renderView = setupRtl(NotificationsContents, {
 });
 
 describe('NotificationsContents', () => {
+  it('marks visible notifications as read when rendered', () => {
+    const notifications = [
+      createStubNotification({ id: '1', unread: true }),
+      createStubNotification({ id: '2', unread: false }),
+      createStubNotification({ id: '3', unread: true }),
+      createStubNotification({ id: '4', unread: true }),
+      createStubNotification({ id: '5', unread: true }),
+    ];
+
+    const { props } = renderView({ notifications });
+
+    expect(props.actions.read).toHaveBeenCalledWith([
+      notifications[0],
+      notifications[2],
+    ]);
+  });
+
+  it('marks newly visible notifications as read when re-rendered', () => {
+    const notifications = [
+      createStubNotification({ id: '1', unread: true }),
+      createStubNotification({ id: '2', unread: false }),
+      createStubNotification({ id: '3', unread: true }),
+      createStubNotification({ id: '4', unread: true }),
+      createStubNotification({ id: '5', unread: true }),
+    ];
+
+    const { props, update } = renderView({ notifications });
+
+    update({
+      notifications: notifications.slice(1),
+    });
+
+    expect(props.actions.read).toHaveBeenCalledWith([
+      notifications[2],
+      notifications[3],
+    ]);
+  });
+
   it('tracks a click for a notification when the notification is clicked', () => {
     const notifications = [
       createStubNotification({ id: '1' }),
@@ -39,10 +77,10 @@ describe('NotificationsContents', () => {
 
     userEvent.click(view.getAllByRole('link')[1]);
 
-    expect(props.actions.read).toHaveBeenCalledWith(notifications[1]);
+    expect(props.actions.read).toHaveBeenCalledWith([notifications[1]]);
   });
 
-  it('does not the notification as read when a read notification is clicked', () => {
+  it('does not mark the notification as read when a read notification is clicked', () => {
     const notifications = [
       createStubNotification({ id: '1' }),
       createStubNotification({ id: '2', unread: false }),